Ported from godot4: Add support for emitting a signal when entering a NavLink - DarkKilauea https://github.com/godotengine/godot/commit/5d8ba2b2d11fc6c4debdf21fa91bddefaa6f3d6d

+++ servers/navigation/navigation_utilities.h
@@ -85,14 +85,14 @@ struct PathQueryResult {
 	Vector<Vector3> path;
 	Vector<int> path_types;
 	Array path_rids;
-	Vector<int> path_owner_ids;
+	Vector<uint64_t> path_owner_ids;
 };
 
 struct PathQueryResult2D {
 	Vector<Vector2> path;
 	Vector<int> path_types;
 	Array path_rids;
-	Vector<int> path_owner_ids;
+	Vector<uint64_t> path_owner_ids;
 };
 
 } //namespace NavigationUtilities
